Move lex_flags codegen to module
