+++ lrlex/src/lib/codegen.rs
@@ -6,7 +6,8 @@ use cfgrammar::{
 use lrpar::LexerTypes;
 
 use crate::{LRNonStreamingLexerDef, LexBuildError, LexFlags, LexerKind};
-use proc_macro2::Ident;
+use proc_macro2::{Ident, TokenStream};
+use quote::{ToTokens, TokenStreamExt, quote};
 use std::{collections::HashMap, fmt, marker::PhantomData, path::Path};
 
 pub(crate) enum LexerSrcEnvError {
@@ -293,5 +294,68 @@ where
                 error: e,
             }
         })
+    }
+
+    pub(crate) fn gen_lex_flags_decl(&self, build_env: &LexerBuildEnv<LexerTypesT>) -> TokenStream {
+        let LexFlags {
+            allow_wholeline_comments,
+            dot_matches_new_line,
+            multi_line,
+            octal,
+            posix_escapes,
+            case_insensitive,
+            unicode,
+            swap_greed,
+            ignore_whitespace,
+            size_limit,
+            dfa_size_limit,
+            nest_limit,
+        } = build_env.lex_flags();
+        let allow_wholeline_comments = QuoteOption(allow_wholeline_comments.as_ref());
+        let dot_matches_new_line = QuoteOption(dot_matches_new_line.as_ref());
+        let multi_line = QuoteOption(multi_line.as_ref());
+        let octal = QuoteOption(octal.as_ref());
+        let posix_escapes = QuoteOption(posix_escapes.as_ref());
+        let case_insensitive = QuoteOption(case_insensitive.as_ref());
+        let unicode = QuoteOption(unicode.as_ref());
+        let swap_greed = QuoteOption(swap_greed.as_ref());
+        let ignore_whitespace = QuoteOption(ignore_whitespace.as_ref());
+        let size_limit = QuoteOption(size_limit.as_ref());
+        let dfa_size_limit = QuoteOption(dfa_size_limit.as_ref());
+        let nest_limit = QuoteOption(nest_limit.as_ref());
+
+        // Code gen for the lexerdef() `lex_flags` variable.
+        quote! {
+            let mut lex_flags = ::lrlex::DEFAULT_LEX_FLAGS;
+            lex_flags.allow_wholeline_comments = #allow_wholeline_comments.or(::lrlex::DEFAULT_LEX_FLAGS.allow_wholeline_comments);
+            lex_flags.dot_matches_new_line = #dot_matches_new_line.or(::lrlex::DEFAULT_LEX_FLAGS.dot_matches_new_line);
+            lex_flags.multi_line = #multi_line.or(::lrlex::DEFAULT_LEX_FLAGS.multi_line);
+            lex_flags.octal = #octal.or(::lrlex::DEFAULT_LEX_FLAGS.octal);
+            lex_flags.posix_escapes = #posix_escapes.or(::lrlex::DEFAULT_LEX_FLAGS.posix_escapes);
+            lex_flags.case_insensitive = #case_insensitive.or(::lrlex::DEFAULT_LEX_FLAGS.case_insensitive);
+            lex_flags.unicode = #unicode.or(::lrlex::DEFAULT_LEX_FLAGS.unicode);
+            lex_flags.swap_greed = #swap_greed.or(::lrlex::DEFAULT_LEX_FLAGS.swap_greed);
+            lex_flags.ignore_whitespace = #ignore_whitespace.or(::lrlex::DEFAULT_LEX_FLAGS.ignore_whitespace);
+            lex_flags.size_limit = #size_limit.or(::lrlex::DEFAULT_LEX_FLAGS.size_limit);
+            lex_flags.dfa_size_limit = #dfa_size_limit.or(::lrlex::DEFAULT_LEX_FLAGS.dfa_size_limit);
+            lex_flags.nest_limit = #nest_limit.or(::lrlex::DEFAULT_LEX_FLAGS.nest_limit);
+            let lex_flags = lex_flags;
+        }
+    }
+}
+
+/// The quote impl of `ToTokens` for `Option` prints an empty string for `None`
+/// and the inner value for `Some(inner_value)`.
+///
+/// This wrapper instead emits both `Some` and `None` variants.
+/// See: [quote #20](https://github.com/dtolnay/quote/issues/20)
+struct QuoteOption<T>(Option<T>);
+
+impl<T: ToTokens> ToTokens for QuoteOption<T> {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        tokens.append_all(match self.0 {
+            Some(ref t) => quote! { ::std::option::Option::Some(#t) },
+            None => quote! { ::std::option::Option::None },
+        });
     }
 }
